requirements.txt
tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
#tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
#tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips
pandas==2.2.0
numpy==1.26.3
opencv-python==4.9.0.80
Pillow

#API
fastapi==0.108.0
uvicorn

streamlit
